apps/discord-reaction-bot/requirements.txt
# Discord bot dependencies
discord.py>=2.3.0

# Async HTTP client for Local AI Router API calls
aiohttp>=3.9.0
